fix(vr): drive held melee by velocity instead of a joint motor
Device report: heavy lag following the hand, and the weapon pivoting about
the wrong point. Both come from the same place.

The six-axis generic-joint position motor was never a sound orientation
servo - Rapier solves AngX/AngY/AngZ independently, which is not shortest-arc
- and the body's origin sits on the weapon *head* rather than in the grip, so
the linear and angular motors torqued and translated each other through that
offset. Measured on a 100-degree swing: 0.14 units of lag, and orientation
error still *growing* thirteen frames after the hand stopped, peaking near 88
degrees. A body that lags rotates about its origin, and its origin is the
weapon head - which is exactly what "the pivot point seems wrong" looks like.

So the pivot is not a separate fix. Each step the weapon is now given the
velocity that lands it on the pose the hand asked for; converging in one step
leaves no residual error to pivot around. It stays dynamic, so world geometry
still stops it dead, actors still take contacts without being launched, and
every contact is still real.

One subtlety worth keeping: the drive reads the target's next_position, not
position. The hand pose arrives via set_next_kinematic_position, which Rapier
commits during the step, so reading the committed pose drives toward where
the hand was last frame - a permanent one-frame lag no gain removes (0.068
units through the swing, against 0.002 reading the pending pose).

melee_spring/melee_damping are gone; melee_max_speed/melee_max_turn replace
them as safety clamps for discontinuous poses, not as feel knobs. The harness
becomes physics::held_melee_drive, and its guard now runs in the ordinary
suite - with no tuning grid there is no global dev-param mutation to isolate.
